Fix gitignore and add serialization runtime files.
the previous `BUILD` is incorrect. ref: https://www.atlassian.com/git/tutorials/saving-changes/gitignore
"Due to a performance-related quirk in Git, you can not negate a file that is ignored due to a pattern matching a directory"
So, rather than ignoring BUILD, we inore all files in BUILD (and then un-ignore those we care about).